+++ src/main/gen/wgslplugin/language/parser/WgslParserUtil.java
@@ -1,0 +1,6 @@
+package wgslplugin.language.parser;
+
+import com.intellij.lang.parser.GeneratedParserUtilBase;
+
+public class WgslParserUtil extends GeneratedParserUtilBase {
+}